migration / dashboard update

+++ Server/DTOs/AppliedExerciseCreateDto.cs
@@ -5,7 +5,7 @@ namespace FitnessTracker.Server.DTOs
         public int ExerciseTemplateId { get; set; } // hareket
         public int Sets { get; set; }
         public int Reps { get; set; }
-        public int Weight { get; set; }
+        public decimal Weight { get; set; }
         public string? Notes { get; set; }
     }
 }
